Fix the text width calculation in case of a very small text size.
Attempt to fix issue #195
It definitely fixes the problem for me.
While I can try to apply post processing of the SVG to change the size of the font to something more seemly I'd rather not. The svg in our system come from a DXF to SVG convertor on the server and recalculating potentially everything in every svg is a waste.